Restructuring for cleaner imports
