+++ ultraSoftCrawfish/funcs/__init__.py
@@ -1,0 +1,2 @@
+import sys
+sys.path.append("../..")
