+++ ultraSoftCrawfish/helpers/pdos_helpers.py
@@ -1,7 +1,6 @@
-from helpers.ase_helpers import get_atoms
-from helpers.data_parsing_helpers import get_kmap_from_atoms, get_el_orb_u_dict
-from helpers.ElecData import parse_data
-from helpers.misc_helpers import get_orb_bool_func
+from ultraSoftCrawfish.helpers.data_parsing_helpers import get_el_orb_u_dict
+from ultraSoftCrawfish.helpers.ElecData import parse_data
+from ultraSoftCrawfish.helpers.misc_helpers import get_orb_bool_func
 import numpy as np
